+++ src/components/MovieForm.tsx
@@ -1,0 +1,145 @@
+import React, { useState } from "react";
+import type { Theme } from "../styles/theme";
+import type { Movie } from "../types/movie";
+
+interface MovieFormProps {
+  currentTheme: Theme;
+  onAddMovie: (movie: Omit<Movie, "id">) => Promise<void>;
+  isLoading: boolean;
+}
+
+export function MovieForm({ currentTheme, onAddMovie, isLoading }: MovieFormProps) {
+  const [newTitle, setNewTitle] = useState<string>("");
+  const [newDirector, setNewDirector] = useState<string>("");
+  const [newYear, setNewYear] = useState<number | "">("");
+  const [newGenre, setNewGenre] = useState<string>("");
+  const [newRating, setNewRating] = useState<number | "">("");
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!newTitle || !newDirector || !newYear || !newGenre || !newRating) {
+      return;
+    }
+
+    const newMovie = {
+      title: newTitle,
+      director: newDirector,
+      year: Number(newYear),
+      genre: newGenre,
+      rating: Number(newRating),
+    };
+
+    await onAddMovie(newMovie);
+    
+    // 성공적으로 추가되면 폼 초기화
+    setNewTitle("");
+    setNewDirector("");
+    setNewYear("");
+    setNewGenre("");
+    setNewRating("");
+  };
+
+  return (
+    <div
+      style={{
+        marginBottom: "24px",
+        padding: "20px",
+        borderRadius: "12px",
+        backgroundColor: currentTheme.componentBg,
+        border: `1px solid ${currentTheme.border}`,
+      }}
+    >
+      <h2>영화 추가</h2>
+      <form
+        onSubmit={handleSubmit}
+        style={{ display: "flex", gap: "12px", flexWrap: "wrap" }}
+      >
+        <input
+          type="text"
+          placeholder="제목"
+          value={newTitle}
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) => setNewTitle(e.target.value)}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+          }}
+        />
+        <input
+          type="text"
+          placeholder="감독"
+          value={newDirector}
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) => setNewDirector(e.target.value)}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+          }}
+        />
+        <input
+          type="number"
+          placeholder="연도"
+          value={newYear}
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) => setNewYear(Number(e.target.value))}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+            width: "80px",
+          }}
+        />
+        <input
+          type="text"
+          placeholder="장르"
+          value={newGenre}
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) => setNewGenre(e.target.value)}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+          }}
+        />
+        <input
+          type="number"
+          placeholder="평점"
+          value={newRating}
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+            const val = Number(e.target.value);
+            if (val >= 0 && val <= 10) setNewRating(val);
+          }}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+            width: "100px",
+          }}
+        />
+        <button
+          type="submit"
+          disabled={isLoading}
+          style={{
+            padding: "8px 16px",
+            backgroundColor: currentTheme.buttonBg,
+            color: currentTheme.buttonText,
+            border: "none",
+            borderRadius: "8px",
+            cursor: isLoading ? "not-allowed" : "pointer",
+            opacity: isLoading ? 0.6 : 1,
+          }}
+        >
+          {isLoading ? "추가 중..." : "추가"}
+        </button>
+      </form>
+    </div>
+  );
+}
